refactor(sanity): update the imports
